refactor: move common properties for system v2 operational settings
